pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.122.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
]
